Revert "Always pull the invocation image (#1723)"
This reverts commit 19a26751930e6202bb9e34ef7cd4bbe21de5ff71.

The change is still something we want eventually but it has exposed
problems elsewhere in our codebase that are hard to track down. Namely
that when we run the integration tests all at once (not just one at a
time manually), the test changes behavior and tries to pull when it
normally wouldn't when run separately.

This causes problems because some of our bundles do exist with the same
reference in dockerhub, but they have a different definition. So pulling
causes all sorts of test failures.

While I track all of this down, I'm going to revert this change until I
can get all the bugs worked out with it.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>