Include calculated parameters in the json/yaml output
When we output an installation with porter installation show, the data
available in the plaintext version should be comparable to what's in the
json/yaml output. Now that we are displaying all resolved parameters,
but only storing the user-provided parameters on the installation
record, we need to output more than just the installation record for
those output formats.

I have changed DisplayInstallation to better wrap an Installation
record, and then put all calculated fields into _calculated so that
people can easily tell which fields are associated with the record (and
can be changed when imported) and which ones are informational only.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>