Call reconcile during porter installation apply
Reconcile the imported changes against the current state of the
installation, and call the appropriate command to match the desired
state. Right now we only trigger under the following conditions:

* different bundle digest
* different resolved parameters
* different credential set _names_ (I don't compare resolved
credentials)

I've added a --force flag to apply so that you can fix stuff when porter
detects no changes but things are borked in reality.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>