Create an installation record for dependencies
When I refactored, it causes a bunch of errors when running a bundle
with dependencies because I wasn't creating an explicit installation
record for the dependency.

I have updated the code so that an installation is created and labled
with

sh.porter.parentInstallation = installation namespace/name

so that later we can filter them out if we want to.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>