Properly format numbers parsed from jsonPath outputs
When we parse outputs for jsonPath expressions, numbers are not being
formatted as expected. For example, a number like "123123123" is being
output as "1.23123123e+08".

This is due to how we are parsing the jsonPath. The output is marshaled
into a generic map, where json numbers are represented as float64. We
use fmt.Sprintf to format the matching output value, and Go by default
formats float with %f/%g, neither of which is the desired format.

I've updated our json parsing to use the Number type so that the format
of the captured value matches what was used in the original output.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>