Install the helm3 mixin on CI machine
When we build the example bundles, they rely on some mixins being
installed on the CI machine. In a recent PR, I updated an example to use
the helm3 mixin since the old helm mixin is deprecated. I forgot to
update the mixins installed in CI, so the example mixin build is
failing.

This should fix that.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>