Fix deploy of wordpress example bundle
With the helm3 mixin using --atomic, helm is waiting for the deployment
to be considered complete. However when installing on kind, the default
service type of loadbalancer will never get an ip address assigned. I've
changed the service type to NodePort so that the helm release completes
successfully and doesn't just timeout waiting for an ip that is never
coming.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>