Make log persistence configurable
Add a flag to the bundle actions allowing the user to skip saving a copy
of the logs to Porter's datastore.

porter install --no-logs

Now when they run porter logs -i mybuns, it will return "no logs found".

This is useful in cases where the logs are too large to store.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>