Use structured logs and export traces
Log to a file in ~/.porter/logs/CORRELLATION_ID.json and in colored human readable
output to stderr. When stderr is not a TTY, color is omitted (which is
what we need for piping to a file or running on headless CI servers).

Tracing has been enabled using open-telemetry and can be configured with
the standard OTEL environment variables.

https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/protocol/exporter.md#specify-protocol

This is just a stub of the final tracing and log switch over, that lays
down a framework for converting the rest and instrumenting more
commands. For now we start a trace for the command executed.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>