Reduce how often we hit the database
* Consolidate creating indices into just 3 calls. We can further refine
this in #1782, when we make initializing the db a single event.
* Do not ping on connect. This does mean that the first query will be
slower as it will establish the connection. So let's be aware of that
and not assume the first call just has horrible performance.

I am still looking into why the first call that establishes the
connection has bad performance.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>