Sync operator with porter v1
These are changes necessary to get the Porter Operator updated to work
with the latest version of Porter v1

* Log which config file we loaded, since the operator injects it in
non-standard spot.
* Print the used config on a separate line. If the file didn't have a
newline at the end, the logs were continuing on the same line as the
config output.
* Allow specifying some runtime flags (normally specified on porter
install) during porter apply: runtime-driver and
allow-docker-host-access.
* Use consistent naming for AllowDockerHostAccess. It was confusing that
we used two different terms for it in the code.
* Log the runtime driver used
* Add mage target for building the porter agent and publishing it to the
operator test cluster using the canary-dev tag, e.g.
localhost:5000/porter-agent:canary-dev

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>