Print the applied file in debug mode
When running porter * apply, print the contents of the applied file in
debug mode so that we can troubleshoot issues in the porter operator,
such as we didn't pass in a valid yaml file or converted it from a CRD
improperly.

Always print the name of the applied resource when completed.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>