Add installation instruction to porter explain command (#1773)
* Only store parameter overrides on the installation record

Installation.Parameters should reflect parameter values that the user
selected, similar to --param on the command line. It shouldn't contain
any resolved values, such as from parameter sets as we already track
those on the Installation.ParameterSets field.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>

* Include calculated parameters in the json/yaml output

When we output an installation with porter installation show, the data
available in the plaintext version should be comparable to what's in the
json/yaml output. Now that we are displaying all resolved parameters,
but only storing the user-provided parameters on the installation
record, we need to output more than just the installation record for
those output formats.

I have changed DisplayInstallation to better wrap an Installation
record, and then put all calculated fields into _calculated so that
people can easily tell which fields are associated with the record (and
can be changed when imported) and which ones are informational only.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>

* Fix test database creation

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>

* Create an installation record for dependencies

When I refactored, it causes a bunch of errors when running a bundle
with dependencies because I wasn't creating an explicit installation
record for the dependency.

I have updated the code so that an installation is created and labled
with

sh.porter.parentInstallation = installation namespace/name

so that later we can filter them out if we want to.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>

* Fix how we resolve parameters for dependencies

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>

* Print params/creds before the bundle is run

Print the params and creds passed to a bundle, including their values.
Make sure that sensitive values are censored.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>

* Uncomment missed tests

I had commented these out while refactoring, and forgot to add them back
when I was done.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>

* Fix a bug from merging with the state PR

The porter state file is not always present for
bundles that do not use state. So skip over it when
it's not found as a parameter.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>

* Hide internal outputs from porter installation outputs list

Do not include internal outputs, such as porter-state.tgz when listing
outputs.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>

* Add first time installation instruction to porter explain command

Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>

* Move installation instruction to the last part of the explain output command

Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>

* fix conflict on go.mod

Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>

* Fix using single credential, set incorrect order of param flag, fix test

Signed-off-by: joshuabezaleel <joshua.bezaleel@gmail.com>

Co-authored-by: Carolyn Van Slyck <me@carolynvanslyck.com>