Use an ipv4 network for test cluster and registry
I am suddenly having trouble using ipv6 on my dev machine and after
hours troubleshooting... I just need to be unblocked.

What I am seeing is that as soon as the registry container is connected
to the kind network (which has ipv6 enabled), I can no longer run docker
push localhost:5000/... It times out and is unable to connect. When I
disconnect the container from a ipv6 enabled network, it works fine.

So for now, I'm making a network named porter with ipv6 disabled, and
connecting both kind and the registry to it so that I'm unblocked.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>