Fix build driver experimental flag checks
There are two config settings that control which build driver we use:

* experimental
* build-driver

The experimental feature needs to be enabled in order for the
build-driver setting to take effect. When I was testing another PR, I
realized that in some parts of the code, I was only checking the
build-driver setting without making sure the experimental feature was
enabled.

This adds a function GetBuildDriver to Config that we can use to
centralize that logic.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>