Use a distroless base image (#1656)
* Use distroless base image

Use a distroless base image for our porter docker images. This
has less of an attack surface because it only ships
the essentials to run porter, not the extra stuff that usually comes
with a linux distribution.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Apply suggestions from code review

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

Co-authored-by: Nathaniel "Church" Hatfield <church13halo@gmail.com>

* Move agent's run.sh into Go

Since the nonroot distroless image doesn't have a shell, we can't use
run.sh to copy the porter config files into PORTER_HOME at container
start. I have implemented that in Go (sorry it's a lot vs what good ole
cp did for us under the hood).

One trick is that when /porter-config is mounted into the container by
k8s, it uses symlinks like this:

/porter-config
  ..data/porter.config
  porter.config -> ..data/porter.config

So it's not a straightforward as you'd think at first glance.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Apply suggestions from code review

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Co-authored-by: Vaughn Dice <vaughn.dice@fermyon.com>
Co-authored-by: Nathaniel "Church" Hatfield <church13halo@gmail.com>
Co-authored-by: Vaughn Dice <vaughn.dice@fermyon.com>