Support uninstall through the installation apply command
Allow someone to manage the entire installation lifecycle with a file,
and not have to jump back to using uninstall. That causes problems when
people are scripting porter calls based on what's in a repo, for
example, they need to detect deleted files and treat them differently.
Also if the bundle is already uninstalled, calling uninstall twice
results in an error, whereas apply does the right thing.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>