Only include required install params in explain command
When porter explain prints a suggested installation command, it should
only include parameters that are required _and_ apply to the install
command. Parameters can be limited to only apply to a particular action
and if it doesn't apply to install, we don't want to tell people to
specify in for porter install.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>