Add templating to porter config file
Support templating in the Porter config file, e.g.
~/.porter/config.json|toml|yaml.

The template syntax uses a yaml-friendly delimiiter, ${}. For example,
${env.NAME} or ${secret.NAME}. If you are using toml or json for your
config file format, then you may need to use quotes around template
values to ensure that the config file has valid syntax when it is
loaded.

We load the config file first with only environment variables
substituted, then we initialize the secrets plugin, and finally do a
second pass of the config file to replace any secrets. So that first
pass needs to be a valid file based on the current format.

For example, in a porter.toml file, secrets should be quoted:

```toml
[[storage]]
  name = "dev"
  plugin = "mongodb"

  [storage.config]
    url = "${secret.connection-string}"
```

Only environment variables and secrets can be substituted. The secrets
are resolved from the default secret storage.

I am implementing this with the liquid template engine (from shopify
that was ported to Go). I don't want us to commit to supporting liquid
templates at this time, it's just an implementaiton detail that liquid
is used at the moment and it could change later.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>