Add schemaVersion to porter.yaml
This adds a document schema version to the Porter manifest (porter.yaml)
similar to what we already have on credential and parameter sets.
The schema of the any document should not change within a major
version.

I've added a page that lists which schema versions are used by various
porter versions.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>