Only push the current invocation image tag
This is a bug from our very first POC for porter. ❤️

When we push the invocation image for a bundle, instead of just pushing
the single tag that is used by the bundle, it is pushing every single
tag available for that image in the docker image cache.

This is why we were seeing increasingly worse performance when
publishing a bundle the more releases you did for it. I didn't notice
there was a problem until I bought a dedicated drive for my docker cache
and suddenly was keeping around  months of temporary builds of the
porter operator bundle and publish started taking 5+ minutes...

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>