Move created/modified to status for params/creds
We've decided to move any calculated fields to the status struct on our
documents. That way it's clear what can be managed by the user and what
is set by Porter.

As part of this I've updated the schema version from 1.0.0 to 1.0.1. If
I could go back in time, I would have used 1.0.0-alpha.1 or something
for the schema changes during prerelease but it's a bit late for that
now, so I'm just bumping the patch to indicate that the schema is
different.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>