Build invocation images that run as nonroot user
As a general security principal, and also to get porter approved for
inclusion in IronBank, we switched our Porter Agent to run as a nonroot
user. What I didn't anticipate is that this effectively requires the
invocation image (installer) to also run as the same user so that they
can share files when run by the Porter Operator.

This updates porter build to run the bundle as the same nonroot user
(65532) and ensure that all files in the bundle (either during build or
injected before running the container) are owned by the nonroot user.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>