Switch from a fork of cnab-go to main
I originally thought that the concept of an installation document would
be included in the cnab spec and Porter was working off of a fork that
included those changes.

Now that we have taken a different direction with the CNAB dependency
spec that doesn't require the runtime to search existing installations,
that isn't necessary anymore.

This gets Porter off of the fork and back on the main branch of cnab-go.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>