Fix assign to project github action
I have updated our github action that assigns issues to the organization
project to use the latest version of the github action. It also now
automatically assigns pull requests too so that we can track what's in
progress.

It uses a github personal access token for our porter bot to get
permission to assign to the project since the default github action
token can't assign to an org project.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>