Fix close issues workflow
When we first started using the github action that closes issues that
were merged into the v1 branch, github didn't have the
pull_request_target workflow type. When you use just the regular
pull_request type, and the PR came from a fork, the secrets.GITHUB_TOKEN
is read only and cannot close an issue.

The new pull_request_target type works with forks.

This fixes errors like

403 Resource not accessible by integration

which required me to manually close issues when the PR was from a fork.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>