Prevent bad config from causing trouble
When we run certain commands, such as help or version, we shouldn't even
bother trying to load configuration. These commands should always
execute quickly and never return an error.

I've added an annotation to commands that shouldn't load config:
version, docs, porter and the help command/flag. When this annotation is
set on a command, we skip loading config.

I have also set a timeout for attempting to connect to a trace exporter
and fixed the error handling so that we gracefully continue to execute
and only log a warning about the trace problem.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>