Do not use helm3 mixin in test bundles
We shouldn't rely on an external mixin in our test bundles.
When we make a change like this (nonroot) then we don't have a good way
to use an updated version of the mixin in our tests.

I've updated the test bundle to act the same but not actually try to do
a helm deployment. It's the same difference for our tests,
and ensures we aren't slowed down trying to fix the helm3 mixin just
to get our changes out.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>