Correct how we run as a non-root user
Previously, I tried running specifically as the nonroot user with the
nonroot group. This caused a lot of file permission problems which I
corrected by having cnab-go set the owner of all copied files to the
container user instead of root.

Then we realized that what we have doesn't work for accessing a mounted
docker socket because it doesn't have permission to it, as it's owned by
root.root on macos and that can't be changed. See
https://github.com/docker/for-mac/issues/4755 for details which is still
an open issue 2 years after it was filed.

On linux the docker socket is mounted with root.docker. So we can use
--group-add docker to ensure the container has the docker group, and
therefore can access the socket.

For mac, accessing the docker socket requires two things:
* The user that the container runs as must be in the root group (gid=0)
* We should mount the /var/run/docker.sock.raw instead of the normal
path. This works around a known permission issue.

I need to investigate more on windows to know what workarounds if any
are required.

In order to have the container be in the root group, it either must be
defined that way when useradd is called, or the user id that the
container runs as must not match an existing user id in the container.
Since the distroless nonroot image used by the client+agent doesn't put
the nonroot user in the root group, I am no longer using it. Instead I'm
using the latest tag (nonroot was just a variant of latest anyway). This
image doesn't have a nonroot user defined, so when we run the container
as 65532, it is placed in the root group.

In the invocation image, I stopped defining a nonroot user/group and am
also using USER 65532, and no longer specifying a group so that it runs
as the root group.

In order to support running as any random UID, when files are copied
into the image, I am setting the group to root and giving it the same
permissions as the owner.
Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>