Build invocation images that run as nonroot user (#1930)
* Build invocation images that run as nonroot user

As a general security principal, and also to get porter approved for
inclusion in IronBank, we switched our Porter Agent to run as a nonroot
user. What I didn't anticipate is that this effectively requires the
invocation image (installer) to also run as the same user so that they
can share files when run by the Porter Operator.

This updates porter build to run the bundle as the same nonroot user
(65532) and ensure that all files in the bundle (either during build or
injected before running the container) are owned by the nonroot user.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Do not use helm3 mixin in test bundles

We shouldn't rely on an external mixin in our test bundles.
When we make a change like this (nonroot) then we don't have a good way
to use an updated version of the mixin in our tests.

I've updated the test bundle to act the same but not actually try to do
a helm deployment. It's the same difference for our tests,
and ensures we aren't slowed down trying to fix the helm3 mixin just
to get our changes out.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Doc nonroot in custom dockerfile page

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Add test case for OCIReference

This adds a test case for OCIReference that ensures that Porter can
parse a reference like carolynvs/whalesayd:v0.1.1@sha256:5cca9...

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Correct how we run as a non-root user

Previously, I tried running specifically as the nonroot user with the
nonroot group. This caused a lot of file permission problems which I
corrected by having cnab-go set the owner of all copied files to the
container user instead of root.

Then we realized that what we have doesn't work for accessing a mounted
docker socket because it doesn't have permission to it, as it's owned by
root.root on macos and that can't be changed. See
https://github.com/docker/for-mac/issues/4755 for details which is still
an open issue 2 years after it was filed.

On linux the docker socket is mounted with root.docker. So we can use
--group-add docker to ensure the container has the docker group, and
therefore can access the socket.

For mac, accessing the docker socket requires two things:
* The user that the container runs as must be in the root group (gid=0)
* We should mount the /var/run/docker.sock.raw instead of the normal
path. This works around a known permission issue.

I need to investigate more on windows to know what workarounds if any
are required.

In order to have the container be in the root group, it either must be
defined that way when useradd is called, or the user id that the
container runs as must not match an existing user id in the container.
Since the distroless nonroot image used by the client+agent doesn't put
the nonroot user in the root group, I am no longer using it. Instead I'm
using the latest tag (nonroot was just a variant of latest anyway). This
image doesn't have a nonroot user defined, so when we run the container
as 65532, it is placed in the root group.

In the invocation image, I stopped defining a nonroot user/group and am
also using USER 65532, and no longer specifying a group so that it runs
as the root group.

In order to support running as any random UID, when files are copied
into the image, I am setting the group to root and giving it the same
permissions as the owner.
Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix driver tests

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix client image

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Ensure group has same permissions as user

When Porter creates a file or directory, the group should be given the
same permissions as the user.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Build images with buildkit

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Remove stray nonroot

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Remove file that shouldn't be checked in

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Use a real user in the root group for the invocation image

Some mixins, like the helm3 mixin, need to run commands as the actual
user (helm init) and there needs to be a real user when it's run. So we
can't rely on just using a rando UID with no backing user.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Improve dockerfile template

* Rename UID to BUNDLE_USER and BUNDLE_UID
* Use the consistent syntax of ${VAR} instead of $VAR
* Update custom dockerfile documentation

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Add Dockerfile support to our code highlighting

Update our prism assests to include support for Dockerfiles

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix formatting in build-image.md

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix grammar

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

Co-authored-by: Vaughn Dice <vaughn.dice@fermyon.com>

Co-authored-by: Vaughn Dice <vaughn.dice@fermyon.com>