Only publish to ghcr.io
We have finished vetting using ghcr.io with Porter and it has been more
stable than Docker Hub. So we are going to stop publishing our images
and example bundles to Docker Hub and only publish to ghcr.io

I will leave the old assets up at Docker Hub and they will continue to
be available for the forseeable future.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>