Add schemaVersion to porter.yaml (#1946)
* Add schemaVersion to porter.yaml

This adds a document schema version to the Porter manifest (porter.yaml)
similar to what we already have on credential and parameter sets.
The schema of the any document should not change within a major
version.

I've added a page that lists which schema versions are used by various
porter versions.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Move the schema template into pkg/schema

All of our other file format schemas are in pkg/schema. This moves the
file to the right location and adds a copy instruction in our build
target to copy the template from the schema directory when we build.

Make the schema available at /schema/v1/NAME.schema.json

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fixes

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Only warn about bundles with no schema set

I'm avoiding returning a hard error immediately because then people would upgrade and no existing
bundles would work for them. Maybe in the future we can change this to a hard check when more
people have had a chance to migrate.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Incorporate review feedback

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix typos

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>