Remove porter manifest from the bundle cache
It's important that we should be able to pull a bundle and fall back to
working with it as a vanilla cnab bundle when it was built with an
unsupported porter version. This removes trying to load the manifest
from the bundle cache and each calling command will need to decide how
to gracefully handle the bundle it pulls from the cache (i.e. some
commands have to require a supported version, like build, but others
like install can still work without a manifest).

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>