Check the manifest schema before build-time commands
The following commands should check the manifest schemaVersion before
attempting to work with a manifest:

* lint
* build
* publish

Really all commands chain off of eachother, so as long as we have
consistent error handling when loading a manifest, all the commands will
have the same behavior.

I've added a config flag "schema-check", that indicates how the user
wants schema checks to occur. By default, we check that the schemaVersion
exactly matches the supported schema. The user may change that handling
using schema-check or PORTER_SCHEMA_CHECK (there is no flag). Its mostly
intended to help people move from one version of Porter to another and
migrate files without a hard error that may make that difficult.

Available values for schema-check are:

* exact - the entire version must match
* major - the MAJOR portion of the version must match
* minor - the MAJOR.MINOR portion of the version must match
* none - Just warn but don't stop execution regardless of how poor the
  match is.

In all cases, when the versions don't match exactly, we will always
print a warning. The config setting only controls if the porter command
immediately fails or if the user is allowed to continue past the
warning.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>