Remove Porter.Manifest
The porter struct has always had a Manifest field that is great at
causing nil pointer exceptions because it's essentially a global
variable and not all code paths populate it.

Since we need to support bundles that may not be able to populate a
manifest, i.e. it's an unsupported manifest schema, or a vanilla CNAB
bundle, I've removed it and made any commands that need the manifest to
load it themselves instead of relying on this global.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>