Bump cnab-go, Go, docker, containerd and k8s (on stable) (#1971)
* Bump cnab-go, cnab-to-oci, Go, docker, containerd and k8s

Update to the most recently patched versions of cnab-go (stable fork),
cnab-to-oci,  Go, docker, containerd, and k8s. The majority of these versions are actually resolved from
cnab-go, so we don't need to maintain replace statements for them
anymore! 🎉

I'm staying on Go 1.16 for the go.mod format because go mod tidy can't
resolve the graph when it's set to go 1.17. On release/v1 we are happily
using the latest go.mod format so I'm not concerned that it's a long
term problem. We still build with Go 1.17, it's just the file format
that is impacted.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Bump to cnab-to-oci 0.3.3

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>