Fix setting GitHub action env vars
I was writing to the file without considering that there could already
have been content in the file, and therefore was undoing any env vars
that other parts of the build had set.

This uses a new version of github.com/carolynvs/magex with built-in
support for interacting with the CI system (so I don't mess it up
again).

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>