Update version on example bundles
Bump the version on the example bundles. We should have done this as
soon as we cut a 1.0.0 release of Porter, since it ended up publishing
examples over the top of existing ones that were compatible with v0.38

Well it's too late to fix that, but we can make sure we bump the version
number when we move to ghcr.io. Later we can optionally go back and
republish the bundles from stable if needed (which will always host with
docker hub, we aren't moving the old stable example bundles to ghcr.io)

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>