Bump magefiles to fix integration tests
When I moved our magefile helpers into a separate repository, I had
missed that the integration test helpers were relying on reading files
from the current repository to create a kind cluster.

Now that the magefiles are in a separate repo, that doesn't work
anymore. I've switched to using go:embed in the magefiles helpers and
this bumps our reference to magefiles to pick up the change.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>