Add gh action to scan porter agent with trivy
If we use the Trivy GitHub action then we can get integrated security
reports sent to GitHub. So I decided to build the porter agent image in
GH actions so we can use trivy there.

This means that if the security scan fails, it won't block us from
cutting a release of Porter.

Long term, it may make sense to move more of our builds to GH actions.
This can be the first step.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>