Fix website BASEURL
I had made a change earlier to netlify.toml and forgot to use export
when setting BASEURL. So at the moment when you go to
release-v1.porter.sh or a pull request preview, the navigation links at
the top all redirect back to porter.sh instead of using the unique
BASEURL specific for that PR/branch.

I have moved the BASEURL logic into mage so that it's not as easy to
break in the future (hopefully).

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>