Add create a bundle task documentation
This adds a task documentation page that explains how to create a
bundle.

I'm still settling on what the organization of the nav vs. the
directory structure of the docs should be. At the moment I'm thinking
that we shouldn't use a directory structure based on possibly changing
characteristics, e.g. the target audience, or the type of documentation
(task, reference, concept) and instead organize by the "noun".

In this case we are discussing bundles, so I put the new page under
bundle/ and reorganized some of the introductory bundle pages that were
previous under the "Create a bundle" navigation structure (the file
format and the manifest reference page).

We really need to reorganize all the docs, but I don't want to let that
large task prevent us from incrementally making process in the meantime.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>