Make runtime manifest log output debug messages
We shouldn't always be printing out log lines when executing the runtime
manifest. Those were always intended to be debug lines and I just missed
gating them with a Debug check.

This makes sure that they are only printed when --debug is specifed and
that they are printed to stderr.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>