Fallback to amd64 on M1 when installing binaries
When the package client (used for mixins and plugins) is looking for a
compatible binary to download, fallback to downloading amd64 for macos
when arm64 is not available. None of the mixins or plugins build to
arm64 yet, so by falling back to amd64 the command succeeds and relies
on the M1 rosetta capabilties to handle running the other format.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>