Add probot settings configuration file
This defines some of our repository configuration in code so that its easier to manage our repositories and keep them consistent. It defines our labels and repository permissions.

The probot settings app updates the repository to apply the settings
in this file when code is pushed to the main branch.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>