Rename claims to installations
In v1 we stopped storing claims directly, which are runtime documents
passed to bundles representing the current action. Instead we now store
documents that better represent the installation, and a run of a bundle,
using terms that we use in Porter, and not CNAB terms that were kind of
misapplied over time as we evolved the data we stored.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>