Move tracer.Close to the RootTraceLogger
I found that it was way too easy to accidentally call span.Close (instead of span.EndSpan)
which then causes the tracer to close. I've moved Close to the RootSpanLogger
and Close now also ends the span. This will prevent someone from closing the tracer when they were
just trying to close a span.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>