Fix plugin load recursion
When porter starts, the main function optionally loads the config file
first, resolving any secrets in the file. When an internal plugin loads,
it should not attempt to load the config file. If it tries, if there are
secrets in the config file, it causes infinite recursion of the secret
plugin loading, trying to resolve secrets by loading another instance of
the secret plugin, and so on.

This fixes the internal plugin run command to skip loading config for
plugins. Plugins get their config passed to them anyway on stdin and
should never try to figure out porter's config file themselves anyway.

I'm also adding a short circuit to the plugin load function to detect
when we are trying to load a plugin from INSIDE another plugin. So that
instead of infinite recursion, it results in an immediate error sent
back to the command output.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>
Co-authored-by: Yingrong Zhao <yingrong.zhao@gmail.com>