Support the interrupt signal SIGINT
When the user presses CTRL+C after running a porter command, we should
attempt to stop as quickly and gracefully as possible.

I have added context.Context in a bunch of places so that any processes
that we start are started with a context. When SIGINT is received, we
cancel the context and the os/exec package handles killing any running
processes.

I've patched our fork of hashicorp/go-plugin to support context
cancellation. I've found that it's still possible to deadlock their
Start and Kill functions because of how locks are used. So the loader
still only gives the plugin only so long to close before intervening so
that we close cleanly.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>