Pin v0.38 install script to compatible package versions
Specify versions of each mixin and plugin to install so that it's
compatible with porter v0.38. I'm still using the same tag for both
porter and exec since they move in lockstep.

Before we cut a v1.0.0 release of porter we need to update this script
to replace latest with the last stable version of v0.38

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>