Do not include plugin config in error
When the plugin configuration can't be marshaled to json, do not print
the plugin config in the error message as it could contain sensitive
data.

I couldn't figure out how to test this since really we are just handling
the error from json.Marshal but there isn't really a way for bad data to
get to that point in Porter. We'd error out much earlier just reading
the config file.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>