Add NoopFeature flag for testing
When I removed the last experimental feature flag, that caused a bunch
of tests to not compile that verify that our experimental feature flags
are working.

I've added a NoopFeature flag that we can use to always have at least
one flag for testing. It's not used in the codebase to enable any
functionality.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>