Add setting for the console log format
The structure-logs feature flag was being used to indicate not only that
the logs and telemetry config sections should be used, but also that the
format of the logs printed to the console should include the timestamp
and log level.

I've added a property to the logs config section, structured, to replace
that now that we aren't using the feature flag anymore.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>