Correct StorageProtocol.FindOptions
When I was adding tests for the FindInstallations function, I realized that I'd made a
mistake earlier when defining the FindOptions. It had erroneously
included Group (which is only for Aggregate). In addition, our wrapper
around FindOptions wasn't exposing the Select field.

I've also updated it to use bson.D/M for consistency with the other options and ensure we are
always passing types that will work with our gRPC conversion logic that
specifically handles bson.D.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>